add deleteImage func
